feat: First part to use CoreData for storing pins and photos
Add Entities, Attributes and Relationships for Pin and Photo. The Pin have an extension with a computed propertie (coordinate) with the type CLLocationCoordinate2D, generated from his latitude/lonitude values.

The DataController as Singleton stores the persistentContainer and viewContext, configure merging with the backgroundContext, saved the Context with an default interval of 30 seconds. The DataController load the persistentStores with didFinishLaunching and save the viewContext with DidEnterBackground and WillTerminate in the AppDelegate.

The MapViewController, CollectionViewController, ViewModels and Annotation use also now the new Pin and Photo type, but must be improve to work correctly.

TODO: Adjust the Saving, Loading and Presention from the new object types from CoreData and the Network Request